// types.ts
export enum SchoolStatus {
    ACTIVE = 'ACTIVE',
    CAUTION = 'CAUTION',
    SUSPENDED = 'SUSPENDED'
}

export interface School {
    id: string;
    name: string;
    address: string;
    motto?: string;
    domain: string;
    badgeUrl?: string | null;
    status: SchoolStatus;
    cautionMessage: string;
    activeSubjectIds: string[];
}